+++ app/src/pages/prompt/ChatTemplateMessageCard.tsx
